requirements.txt
serpapi
python-dotenv
